pyproject.toml
[project]
name = "prompt-spec"
version = "0.1.4"
description = "Lightweight pydantic wrapper prompting library to support sharing of oncology-related LLM prompts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "linkml>=1.9.5",
    "linkml-runtime>=1.9.5",
    "openai>=2.8.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "ruamel-yaml>=0.18.16",
    "typer>=0.20.0",
]

[project.scripts]
prompt-spec = "prompt_spec.cli:main"
